Header/block inventory for peers
Added an inventory of blocks and headers for each peer. When a peer receives
a header, it will now send it to all of the peers that don't already have
it. This code assumes that UDP won't fail. Eventually, this assumption needs
to be replaced with some TCP code as a backup just for header messages.