Added (buggy) BTMerkleTree class to handle partial tree validation
This commit adds a class to lib/bttrees.py that will allow you to
add unvalidated hashes to the tree one by one. The class will check
to see if the unvalidated hashes can be connected to the merkle root
hash when combined with other hashes in its "purgatory". If not, the
hash will be stored in purgatory. If it can, then it and its sibling
will be added to the validated tree. This class is almost complete
and functional, but I missed one step in the recursion logic in the
addhash method, and I'm sleepy now.

This commit also adds mininode.py from the bitcoind python RPC test
framework, as it has been more recently updated than the halfnode.py
and bitcoinnode.py versions that are currently in this tree. Those
two files need to be removed in a subsequent commit.